replica: remove RocksDB and the migrated legacy database
Remove the last of the RocksDB machinery from the replica. The
grocksdb dependency, the Makefile install/build targets, the RocksDB
install steps in replica/README.md, and the RocksDB build in
docker/Dockerfile.alpine-base are all gone; the replica now builds as
a standard pure-Go component.

On boot the replica now deletes the legacy <DataDir>/replica.db once
the migration marker is present, freeing the disk space that the
migrated RocksDB database still occupied. The marker is also recorded
on fresh installs, so every boot path reaches the deletion check.

Upgrading operators MUST pass through 0.0.98 first: that release
migrates existing RocksDB data into the Pebble databases and records
the marker. This release then removes replica.db on first boot; if it
finds legacy RocksDB data without the marker it refuses to start
rather than risk silent data loss. After this release boots, the
migrated RocksDB data is gone, and rolling back to a pre-migration
build will not recover it.

(cherry picked from commit 4f6911a212bcacea60d960fe02d43343b2f5cfbd)
