+++ .github/workflows/alpine-base-image.yml
@@ -33,12 +33,10 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v3
 
-      - name: Extract RocksDB version from Dockerfile
+      - name: Extract Go version from Dockerfile
         id: versions
         run: |
-          rocksdb_ver=$(grep 'ARG ROCKSDB_VERSION=' docker/Dockerfile.alpine-base | cut -d= -f2)
           go_ver=$(grep 'ARG GO_VERSION=' docker/Dockerfile.alpine-base | cut -d= -f2)
-          echo "rocksdb=${rocksdb_ver}" >> "$GITHUB_OUTPUT"
           echo "go=${go_ver}" >> "$GITHUB_OUTPUT"
 
       - name: Build and push
@@ -49,6 +47,6 @@ jobs:
           push: true
           tags: |
             ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:latest
-            ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:rocksdb-${{ steps.versions.outputs.rocksdb }}-go${{ steps.versions.outputs.go }}
+            ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:go${{ steps.versions.outputs.go }}
           cache-from: type=gha
           cache-to: type=gha,mode=max
