+++ .github/workflows/govulncheck.yml
@@ -15,7 +15,5 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version-file: go.mod
-      - name: Install RocksDB dependencies
-        run: make install-replica-deps
       - name: Run govulncheck
         run: go run golang.org/x/vuln/cmd/govulncheck@latest ./...
