+++ Makefile
@@ -1,5 +1,5 @@
 
-.PHONY: all test test-unit test-replica bench-replica bench-sphinx bench-handshake test-config sphincsplus clean server dirauth genconfig ping courier echo-plugin fetch genkeypair geometry http-proxy-client http-proxy-server kpclientd map sphinx replica install-replica-deps install-replica-deps-root rocksdb-local replica-local
+.PHONY: all test test-unit test-replica bench-replica bench-sphinx bench-handshake test-config sphincsplus clean server dirauth genconfig ping courier echo-plugin fetch genkeypair geometry http-proxy-client http-proxy-server kpclientd map sphinx replica
 
 .PHONY: update-go-deps
 update-go-deps:
@@ -53,95 +53,8 @@ kpclientd:
 sphinx:
 	cd cmd/sphinx; go build -trimpath -ldflags "-s -w"
 
-# Privilege escalation for the system-wide RocksDB install. Defaults to sudo;
-# install-replica-deps-root clears it to run the recipe as root directly.
-SUDO ?= sudo
-
-ROCKSDB_VERSION = 10.2.1
-# Per-user RocksDB prefix used by the rocksdb-local / replica-local targets.
-ROCKSDB_LOCAL_PREFIX ?= $(HOME)/.cache/katzenpost/rocksdb-$(ROCKSDB_VERSION)
-
-install-replica-deps:
-	@set -e; \
-	echo "Checking for RocksDB $(ROCKSDB_VERSION)..."; \
-	installed_ver="$$(pkg-config --modversion rocksdb 2>/dev/null || echo none)"; \
-	if [ "$$installed_ver" = "$(ROCKSDB_VERSION)" ]; then \
-		echo "RocksDB $(ROCKSDB_VERSION) already installed"; \
-	else \
-		if [ "$$installed_ver" != "none" ]; then \
-			echo "RocksDB $$installed_ver found, need $(ROCKSDB_VERSION) — removing old version..."; \
-			$(SUDO) rm -f /usr/local/lib/librocksdb.*; \
-			$(SUDO) rm -rf /usr/local/include/rocksdb; \
-			$(SUDO) rm -f /usr/local/lib/pkgconfig/rocksdb.pc; \
-			$(SUDO) ldconfig; \
-		else \
-			echo "RocksDB not found"; \
-		fi; \
-		echo "Installing build dependencies..."; \
-		$(SUDO) apt-get install -y \
-			cmake build-essential pkg-config gcc-14 g++-14 \
-			libsnappy-dev libzstd-dev liblz4-dev \
-			zlib1g-dev libbz2-dev liburing-dev libgflags-dev; \
-		echo "Building RocksDB $(ROCKSDB_VERSION) from source..."; \
-		tmpdir="$$(mktemp -d)"; \
-		cd "$$tmpdir"; \
-		git clone --depth 1 --branch v$(ROCKSDB_VERSION) https://github.com/facebook/rocksdb.git; \
-		cd rocksdb; \
-		env CC=gcc-14 CXX=g++-14 make shared_lib -j$$(nproc); \
-		echo "Installing RocksDB $(ROCKSDB_VERSION)..."; \
-		$(SUDO) env CC=gcc-14 CXX=g++-14 make install-shared; \
-		$(SUDO) ldconfig; \
-		rm -rf "$$tmpdir"; \
-		echo "RocksDB $(ROCKSDB_VERSION) installed successfully!"; \
-	fi
-
-# Build replica (requires RocksDB dependencies)
-# this may require gcc-14
-replica: install-replica-deps
-	cd cmd/replica; CC=gcc-14 CGO_ENABLE=1 CGO_LDFLAGS="-lrocksdb -lstdc++ -lbz2 -lm -lz -lsnappy -llz4 -lzstd -luring" go build -v -trimpath -ldflags "-X github.com/carlmjohnson/versioninfo.Revision=$$(git rev-parse --short HEAD)"
-
-
-# RocksDB built into a per-user prefix; lets the replica be built without root
-# or sudo. Requires the compression dev libraries (install-replica-deps-root).
-rocksdb-local:
-	@set -e; \
-	prefix="$(ROCKSDB_LOCAL_PREFIX)"; \
-	if [ -e "$$prefix/lib/librocksdb.so" ]; then \
-		echo "RocksDB $(ROCKSDB_VERSION) already present at $$prefix"; \
-		exit 0; \
-	fi; \
-	command -v gcc-14 >/dev/null || { echo "gcc-14 not found; run: make install-replica-deps-root"; exit 1; }; \
-	missing=""; \
-	for lib in snappy zstd lz4 z bz2 uring; do \
-		printf 'int main(void){return 0;}\n' | gcc-14 -x c - -l$$lib -o /dev/null 2>/dev/null || missing="$$missing $$lib"; \
-	done; \
-	if [ -n "$$missing" ]; then \
-		echo "Missing compression libraries:$$missing"; \
-		echo "Install the build dependencies once as root: make install-replica-deps-root"; \
-		exit 1; \
-	fi; \
-	echo "Building RocksDB $(ROCKSDB_VERSION) into $$prefix ..."; \
-	tmpdir="$$(mktemp -d)"; \
-	cd "$$tmpdir"; \
-	git clone --depth 1 --branch v$(ROCKSDB_VERSION) https://github.com/facebook/rocksdb.git; \
-	cd rocksdb; \
-	env CC=gcc-14 CXX=g++-14 make shared_lib -j$$(nproc); \
-	env CC=gcc-14 CXX=g++-14 make install-shared INSTALL_PATH="$$prefix"; \
-	rm -rf "$$tmpdir"; \
-	echo "RocksDB $(ROCKSDB_VERSION) installed at $$prefix"
-
-# Build the replica against the per-user RocksDB prefix; no root, no sudo. The
-# prefix is baked in as an rpath so the binary runs without LD_LIBRARY_PATH.
-replica-local: rocksdb-local
-	cd cmd/replica; CC=gcc-14 CGO_ENABLED=1 \
-		CGO_CFLAGS="-I$(ROCKSDB_LOCAL_PREFIX)/include" \
-		CGO_LDFLAGS="-L$(ROCKSDB_LOCAL_PREFIX)/lib -Wl,-rpath,$(ROCKSDB_LOCAL_PREFIX)/lib -lrocksdb -lstdc++ -lbz2 -lm -lz -lsnappy -llz4 -lzstd -luring" \
-		go build -v -trimpath -ldflags "-X github.com/carlmjohnson/versioninfo.Revision=$$(git rev-parse --short HEAD)"
-
-# Install the replica build dependencies system-wide. Run this once as root
-# (no sudo needed); then build with `make replica` as your normal user.
-install-replica-deps-root:
-	$(MAKE) install-replica-deps SUDO=
+replica:
+	cd cmd/replica; go build -v -trimpath -ldflags "-X github.com/carlmjohnson/versioninfo.Revision=$$(git rev-parse --short HEAD)"
 
 clean:
 	rm -f cmd/server/server cmd/dirauth/dirauth cmd/genconfig/genconfig cmd/ping/ping \
@@ -177,16 +90,16 @@ test-unit: test-config
 
 	@echo "All unit tests completed successfully!"
 
-# Run replica unit tests (requires RocksDB dependencies)
+# Run replica unit tests
 test-replica: test-config
 	@echo "Running replica unit tests..."
-	cd replica && GORACE=history_size=7 CC=gcc-14 CGO_ENABLE=1 CGO_LDFLAGS="-lrocksdb -lstdc++ -lbz2 -lm -lz -lsnappy -llz4 -lzstd -luring" go test -coverprofile=coverage.out -race -v -failfast -timeout 30m ./...
+	cd replica && GORACE=history_size=7 go test -coverprofile=coverage.out -race -v -failfast -timeout 30m ./...
 	@echo "Replica unit tests completed successfully!"
 
-# Run replica benchmarks (requires RocksDB dependencies)
+# Run replica benchmarks
 bench-replica:
 	@echo "Running replica benchmarks..."
-	cd replica && CC=gcc-14 CGO_ENABLE=1 CGO_LDFLAGS="-lrocksdb -lstdc++ -lbz2 -lm -lz -lsnappy -llz4 -lzstd -luring" go test -v -run=^$$ -bench=. -benchtime=3x ./...
+	cd replica && go test -v -run=^$$ -bench=. -benchtime=3x ./...
 	@echo "Replica benchmarks completed successfully!"
 
 # Run all sphinx benchmarks
@@ -221,8 +134,8 @@ bench-handshake:
 	@echo "=== Courier Handshake Benchmarks ==="
 	go test -v -run=^$$ -bench=. -benchtime=3x ./courier/server/
 	@echo ""
-	@echo "=== Replica Handshake Benchmarks (requires RocksDB) ==="
-	cd replica && CC=gcc-14 CGO_ENABLE=1 CGO_LDFLAGS="-lrocksdb -lstdc++ -lbz2 -lm -lz -lsnappy -llz4 -lzstd -luring" go test -v -run=^$$ -bench=. -benchtime=3x ./...
+	@echo "=== Replica Handshake Benchmarks ==="
+	cd replica && go test -v -run=^$$ -bench=. -benchtime=3x ./...
 	@echo ""
 	@echo "All wire handshake benchmarks completed successfully!"
