+++ docker/Dockerfile.alpine-base
@@ -1,32 +1,15 @@
 FROM docker.io/alpine:3.21
 
-ARG ROCKSDB_VERSION=10.2.1
 ARG GO_VERSION=1.26.2
 
 RUN adduser katzenpost --gecos "" --disabled-password
 
-# Install build tools and runtime dependencies (but NOT rocksdb-dev from Alpine)
+# Install build tools. Go needs a C toolchain (gcc + musl-dev) for the
+# cgo components; git is used to fetch Go modules during the docker
+# build.
 RUN apk update && apk upgrade && \
-    apk add gcc musl-dev make pv g++ curl netcat-openbsd \
-        bash perl coreutils git cmake linux-headers \
-        snappy-dev zstd-dev lz4-dev zlib-dev bzip2-dev
-
-# Build and install RocksDB from source at a pinned version
-RUN cd /tmp && \
-    git clone --depth 1 --branch v${ROCKSDB_VERSION} https://github.com/facebook/rocksdb.git && \
-    cd rocksdb && \
-    PORTABLE=1 make shared_lib -j$(nproc) && \
-    make install-shared && \
-    cd / && rm -rf /tmp/rocksdb
-
-# Remove build-only deps while apk still works (before adding /usr/local/lib
-# to the linker path, which would shadow system libz.so and break apk)
-RUN apk del git cmake linux-headers
-
-# Ensure musl linker can find RocksDB in /usr/local/lib.
-# ld-musl-x86_64.path REPLACES the default search path, so we must
-# include the standard directories explicitly.
-RUN echo "/lib:/usr/local/lib:/usr/lib" > /etc/ld-musl-x86_64.path
+    apk add gcc musl-dev make pv curl netcat-openbsd \
+        bash perl coreutils git
 
 # Install Go
 RUN curl -fsSL https://go.dev/dl/go${GO_VERSION}.linux-amd64.tar.gz | tar -C /usr/local -xzf - && \
