+++ replica/state_test.go
@@ -184,7 +184,7 @@ func TestStateClosedDatabaseGuards(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 	st.Close()
 
@@ -217,7 +217,7 @@ func TestStateReadCorruptedValue(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 	defer st.Close()
 
@@ -234,7 +234,7 @@ func TestHandleReplicaWriteRejectsWhenStorageFull(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 	defer st.Close()
 
@@ -250,7 +250,7 @@ func TestRebalanceSkipsMalformedKey(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 	defer st.Close()
 
@@ -266,7 +266,7 @@ func TestDBOnDiskBytes(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 
 	n, ok := st.dbOnDiskBytes()
@@ -284,7 +284,7 @@ func TestRefreshStorageFullQuota(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dataDir)
 
-	st := newMigrateTestState(t, dataDir)
+	st := newTestState(t, dataDir)
 	st.initDB()
 	defer st.Close()
