+++ replica/storage_space_windows.go
@@ -6,10 +6,9 @@
 package replica
 
 // availableBytes has no portable syscall.Statfs equivalent on Windows.
-// The replica is a RocksDB/CGO Linux service in practice; on Windows
-// the filesystem free-space reserve is simply not enforced (false),
-// and the MaxStorageMiB quota still applies. Returning false makes
-// the watcher skip the reserve check rather than guess.
+// On Windows the filesystem free-space reserve is simply not enforced
+// (false), and the MaxStorageMiB quota still applies. Returning false
+// makes the watcher skip the reserve check rather than guess.
 func availableBytes(path string) (uint64, bool) {
 	return 0, false
 }
